+++ db/models/document_party.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey
 from sqlalchemy.orm import relationship
 
 from db.base import Base
@@ -7,8 +7,8 @@ from db.base import Base
 class DocumentParty(Base):
     __tablename__ = "juridical_party_documents"
 
-    id = Column(Integer, primary_key=True)
-    party_id = Column(Integer, ForeignKey("juridical_parties.id", ondelete="CASCADE"), index=True)
+    id = Column(BigInteger, primary_key=True)
+    party_id = Column(BigInteger, ForeignKey("juridical_parties.id", ondelete="CASCADE"), index=True)
     type = Column(String(20), nullable=False)
     value = Column(String(64), nullable=False)
