+++ db/models/movement.py
@@ -1,12 +1,13 @@
-from db.base import Base
-from sqlalchemy import Column, Integer, ForeignKey, DateTime, String, Text
+from sqlalchemy import Column, Integer, BigInteger, ForeignKey, DateTime, String, Text
 from sqlalchemy.orm import relationship
+
+from db.base import Base
 
 
 class Movement(Base):
     __tablename__ = 'juridical_movements'
-    id = Column(Integer, primary_key=True)
-    process_id = Column(Integer, ForeignKey('juridical_processes.id', ondelete='CASCADE'), index=True)
+    id = Column(BigInteger, primary_key=True)
+    process_id = Column(BigInteger, ForeignKey('juridical_processes.id', ondelete='CASCADE'), index=True)
     idx = Column(Integer, nullable=True)
 
     created_at = Column(DateTime(timezone=True), nullable=True)
